+++ CHANGELOG.md
@@ -15,7 +15,10 @@
 * AADCrossTenantAccessPolicyConfigurationPartner
   * Initial release
     FIXES [#3253](https://github.com/microsoft/Microsoft365DSC/issues/3253)
- * IntuneSettingCatalogCustomPolicyWindows10
+* EXOQuarantinePolicy
+  * Support exporting global quarantine policy
+    FIXES [#3285](https://github.com/microsoft/Microsoft365DSC/issues/3285)
+* IntuneSettingCatalogCustomPolicyWindows10
   * Initial release
   FIXES [#2692](https://github.com/microsoft/Microsoft365DSC/issues/2692),
   FIXES [#2976](https://github.com/microsoft/Microsoft365DSC/issues/2976),